Move mock respurces to own package
because jersey reource finder may fail if same package also in some
dependency
